requirements.txt
# ====== База ======
numpy==2.1.3
pandas==2.2.3
matplotlib==3.9.3
packaging==24.2
pillow==11.0.0
python-dateutil==2.9.0.post0
six==1.17.0
typing_extensions==4.12.2
setuptools==75.6.0
tzdata==2024.2

# ====== Gym/Farama ======
Farama-Notifications==0.0.4
gym-notices==0.0.8
# ВАЖНО: фиксируем gymnasium на 0.29.1, т.к. SB3 стабильно работает до 1.0
gymnasium==0.29.1

# ====== Stable-Baselines3 (стабильный релиз под gymnasium 0.29.x) ======
stable-baselines3==2.3.0

# ====== PyTorch CUDA 12.4 (официальные manylinux-колёса) ======
torch==2.5.1
# Triton ПИН НЕ СТАВИМ — пусть ставится совместимая версия автоматически
# (если очень хочется закрепить: triton>=3.2,<3.5)

# ====== NVIDIA runtime libs (опционально; PyTorch уже несёт нужные зависимости)
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvtx-cu12==12.4.127
nvidia-nvjitlink-cu12==12.4.127

# ====== Прочее ======
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.1
fsspec==2024.10.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
pyparsing==3.2.0
sympy==1.13.1
